requirements.txt
PyQt5>=5.15.4
numpy>=1.19.5
Pillow>=8.3.1
opencv-python>=4.5.4.60
matplotlib>=3.4.1
setuptools>=56.0.0
aioconsole>=0.6.1
pyserial~=3.5
pypylon~=1.8.0
pymavlink~=2.4.40
numba~=0.58.1
pip~=21.3.1
wheel~=0.37.1
lxml~=4.9.2
future~=0.18.3
Flask~=2.3.2